reorganize registry a bit
